[add] graphite client script to tail and ship data to graphite (#1)
The graphite-client script expects the bpftrace to output data to a log somewhere to be consumed, captured, and shipped. I employed a loosely coupled approach to achieving the desired result, this will ensure that both scripts can run or break independently without interfering with the other. I recommend using a process control system like supervisord to ensure that both scripts are continuously running in addition to configuring log rotate to clean up the logs - this will ensure that the script can run non-interactively.